// src/types/generated/models.ts
export type Access = {
    data?: AccessData
    delete: boolean
    externalize: boolean
    manage: boolean
    read: boolean
    update: boolean
    write: boolean
}

export type AccessData = {
    read: boolean
    write: boolean
}

export type AddOperation = {
    op: string
    path: string
    value: Record<string, any>
}

export type AggregateDataExchange = {
    access: Access
    attributeValues: Array<AttributeValue>
    code: string
    created: string
    createdBy: User
    displayName: string
    favorite: boolean
    favorites: Array<string>
    href: string
    id: string
    lastUpdated: string
    lastUpdatedBy: User
    name: string
    sharing: Sharing
    source: Source
    target: Target
    translations: Array<Translation>
    user: User
}
// ...
    export enum type {
        PERSONAL_ACCESS_TOKEN_V1 = 'PERSONAL_ACCESS_TOKEN_V1',
    }
// ...
export type Attribute = {
    access: Access
    attributeValues: Array<AttributeValue>
    categoryAttribute: boolean
    categoryOptionAttribute: boolean
    categoryOptionComboAttribute: boolean
    categoryOptionGroupAttribute: boolean
    categoryOptionGroupSetAttribute: boolean
    code: string
    constantAttribute: boolean
    created: string
    createdBy: User
    dataElementAttribute: boolean
    dataElementGroupAttribute: boolean
    dataElementGroupSetAttribute: boolean
    dataSetAttribute: boolean
    description: string
    displayDescription: string
    displayFormName: string
    displayName: string
    displayShortName: string
    documentAttribute: boolean
    eventChartAttribute: boolean
    eventReportAttribute: boolean
    favorite: boolean
    favorites: Array<string>
    formName: string
    href: string
    id: string
    indicatorAttribute: boolean
    indicatorGroupAttribute: boolean
    lastUpdated: string
    lastUpdatedBy: User
    legendSetAttribute: boolean
    mandatory: boolean
    mapAttribute: boolean
    name: string
    objectTypes: Array<string>
    optionAttribute: boolean
    optionSet: OptionSet
    optionSetAttribute: boolean
    organisationUnitAttribute: boolean
    organisationUnitGroupAttribute: boolean
    organisationUnitGroupSetAttribute: boolean
    programAttribute: boolean
    programIndicatorAttribute: boolean
    programStageAttribute: boolean
    relationshipTypeAttribute: boolean
    sectionAttribute: boolean
    sharing: Sharing
    shortName: string
    sortOrder: number
    sqlViewAttribute: boolean
    trackedEntityAttributeAttribute: boolean
    trackedEntityTypeAttribute: boolean
    translations: Array<Translation>
    unique: boolean
    user: User
    userAttribute: boolean
    userGroupAttribute: boolean
    validationRuleAttribute: boolean
    validationRuleGroupAttribute: boolean
    valueType: Attribute.valueType
    visualizationAttribute: boolean
}

export namespace Attribute {
    export enum valueType {
        TEXT = 'TEXT',
        LONG_TEXT = 'LONG_TEXT',
        MULTI_TEXT = 'MULTI_TEXT',
        LETTER = 'LETTER',
        PHONE_NUMBER = 'PHONE_NUMBER',
        EMAIL = 'EMAIL',
        BOOLEAN = 'BOOLEAN',
        TRUE_ONLY = 'TRUE_ONLY',
        DATE = 'DATE',
        DATETIME = 'DATETIME',
        TIME = 'TIME',
        NUMBER = 'NUMBER',
        UNIT_INTERVAL = 'UNIT_INTERVAL',
        PERCENTAGE = 'PERCENTAGE',
        INTEGER = 'INTEGER',
        INTEGER_POSITIVE = 'INTEGER_POSITIVE',
        INTEGER_NEGATIVE = 'INTEGER_NEGATIVE',
        INTEGER_ZERO_OR_POSITIVE = 'INTEGER_ZERO_OR_POSITIVE',
        TRACKER_ASSOCIATE = 'TRACKER_ASSOCIATE',
        USERNAME = 'USERNAME',
        COORDINATE = 'COORDINATE',
        ORGANISATION_UNIT = 'ORGANISATION_UNIT',
        REFERENCE = 'REFERENCE',
        AGE = 'AGE',
        URL = 'URL',
        FILE_RESOURCE = 'FILE_RESOURCE',
        IMAGE = 'IMAGE',
        GEOJSON = 'GEOJSON',
    }
}

export type AttributeValue = {
    attribute: Attribute
    value: string
}
// ...
export type Category = {
    access: Access
    aggregationType: Category.aggregationType
    allItems: boolean
    attributeValues: Array<AttributeValue>
    categoryCombos: Array<CategoryCombo>
    categoryOptions: Array<CategoryOption>
    code: string
    created: string
    createdBy: User
    dataDimension: boolean
    dataDimensionType: Category.dataDimensionType
    description: string
    dimension: string
    dimensionItemKeywords: DimensionItemKeywords
    displayDescription: string
    displayFormName: string
    displayName: string
    displayShortName: string
    favorite: boolean
    favorites: Array<string>
    filter: string
    formName: string
    href: string
    id: string
    items: Array<DimensionalItemObject>
    lastUpdated: string
    lastUpdatedBy: User
    legendSet: LegendSet
    name: string
    optionSet: OptionSet
    programStage: ProgramStage
    repetition: EventRepetition
    sharing: Sharing
    shortName: string
    translations: Array<Translation>
    user: User
    valueType: Category.valueType
}

export namespace Category {
    export enum aggregationType {
        SUM = 'SUM',
        AVERAGE = 'AVERAGE',
        AVERAGE_SUM_ORG_UNIT = 'AVERAGE_SUM_ORG_UNIT',
        LAST = 'LAST',
        LAST_AVERAGE_ORG_UNIT = 'LAST_AVERAGE_ORG_UNIT',
        LAST_LAST_ORG_UNIT = 'LAST_LAST_ORG_UNIT',
        LAST_IN_PERIOD = 'LAST_IN_PERIOD',
        LAST_IN_PERIOD_AVERAGE_ORG_UNIT = 'LAST_IN_PERIOD_AVERAGE_ORG_UNIT',
        FIRST = 'FIRST',
        FIRST_AVERAGE_ORG_UNIT = 'FIRST_AVERAGE_ORG_UNIT',
        FIRST_FIRST_ORG_UNIT = 'FIRST_FIRST_ORG_UNIT',
        COUNT = 'COUNT',
        STDDEV = 'STDDEV',
        VARIANCE = 'VARIANCE',
        MIN = 'MIN',
        MAX = 'MAX',
        MIN_SUM_ORG_UNIT = 'MIN_SUM_ORG_UNIT',
        MAX_SUM_ORG_UNIT = 'MAX_SUM_ORG_UNIT',
        NONE = 'NONE',
        CUSTOM = 'CUSTOM',
        DEFAULT = 'DEFAULT',
    }

    export enum dataDimensionType {
        DISAGGREGATION = 'DISAGGREGATION',
        ATTRIBUTE = 'ATTRIBUTE',
    }

    export enum valueType {
        TEXT = 'TEXT',
        LONG_TEXT = 'LONG_TEXT',
        MULTI_TEXT = 'MULTI_TEXT',
        LETTER = 'LETTER',
        PHONE_NUMBER = 'PHONE_NUMBER',
        EMAIL = 'EMAIL',
        BOOLEAN = 'BOOLEAN',
        TRUE_ONLY = 'TRUE_ONLY',
        DATE = 'DATE',
        DATETIME = 'DATETIME',
        TIME = 'TIME',
        NUMBER = 'NUMBER',
        UNIT_INTERVAL = 'UNIT_INTERVAL',
        PERCENTAGE = 'PERCENTAGE',
        INTEGER = 'INTEGER',
        INTEGER_POSITIVE = 'INTEGER_POSITIVE',
        INTEGER_NEGATIVE = 'INTEGER_NEGATIVE',
        INTEGER_ZERO_OR_POSITIVE = 'INTEGER_ZERO_OR_POSITIVE',
        TRACKER_ASSOCIATE = 'TRACKER_ASSOCIATE',
        USERNAME = 'USERNAME',
        COORDINATE = 'COORDINATE',
        ORGANISATION_UNIT = 'ORGANISATION_UNIT',
        REFERENCE = 'REFERENCE',
        AGE = 'AGE',
        URL = 'URL',
        FILE_RESOURCE = 'FILE_RESOURCE',
        IMAGE = 'IMAGE',
        GEOJSON = 'GEOJSON',
    }
}
// ...
export type CategoryOptionGroupSet = {
    access: Access
    aggregationType: CategoryOptionGroupSet.aggregationType
    allItems: boolean
    attributeValues: Array<AttributeValue>
    categoryOptionGroups: Array<CategoryOptionGroup>
    code: string
    created: string
    createdBy: User
    dataDimension: boolean
    dataDimensionType: CategoryOptionGroupSet.dataDimensionType
    description: string
    dimension: string
    dimensionItemKeywords: DimensionItemKeywords
    displayDescription: string
    displayFormName: string
    displayName: string
    displayShortName: string
    favorite: boolean
    favorites: Array<string>
    filter: string
    formName: string
    href: string
    id: string
    items: Array<DimensionalItemObject>
    lastUpdated: string
    lastUpdatedBy: User
    legendSet: LegendSet
    name: string
    optionSet: OptionSet
    programStage: ProgramStage
    repetition: EventRepetition
    sharing: Sharing
    shortName: string
    translations: Array<Translation>
    user: User
    valueType: CategoryOptionGroupSet.valueType
}
// ...
    export enum trackerType {
        TRACKED_ENTITY = 'TRACKED_ENTITY',
        ENROLLMENT = 'ENROLLMENT',
        EVENT = 'EVENT',
        RELATIONSHIP = 'RELATIONSHIP',
    }
// ...
export type Error = {
    errorCode: string
    message: string
    trackerType: string
    uid: string
}
// ...
export type FileResource = {
    access: Access
    attributeValues: Array<AttributeValue>
    code: string
    contentLength: number
    contentMd5: string
    contentType: string
    created: string
    createdBy: User
    displayName: string
    domain: FileResource.domain
    favorite: boolean
    favorites: Array<string>
    hasMultipleStorageFiles: boolean
    href: string
    id: string
    lastUpdated: string
    lastUpdatedBy: User
    name: string
    sharing: Sharing
    storageStatus: FileResource.storageStatus
    translations: Array<Translation>
    user: User
}

export namespace FileResource {
    export enum domain {
        DATA_VALUE = 'DATA_VALUE',
        PUSH_ANALYSIS = 'PUSH_ANALYSIS',
        DOCUMENT = 'DOCUMENT',
        MESSAGE_ATTACHMENT = 'MESSAGE_ATTACHMENT',
        USER_AVATAR = 'USER_AVATAR',
        ORG_UNIT = 'ORG_UNIT',
        CUSTOM_ICON = 'CUSTOM_ICON',
    }

    export enum storageStatus {
        NONE = 'NONE',
        PENDING = 'PENDING',
        FAILED = 'FAILED',
        STORED = 'STORED',
    }
}
// ...
export type Sharing = {
    external: boolean
    owner: string
    public: string
    userGroups: Record<string, UserGroupAccess>
    users: Record<string, UserAccess>
}
// ...
export type Source = {
    params: SourceParams
    requests: Array<SourceRequest>
}
// ...
export type Target = {
    api: Api
    request: TargetRequest
    type: Target.type
}

export namespace Target {
    export enum type {
        INTERNAL = 'INTERNAL',
        EXTERNAL = 'EXTERNAL',
    }
}
// ...
export type Translation = {
    locale: string
    property: string
    value: string
}
// ...
export type User = {
    access: Access
    accountExpiry: string
    attributeValues: Array<AttributeValue>
    avatar: FileResource
    birthday: string
    catDimensionConstraints: Array<Category>
    code: string
    cogsDimensionConstraints: Array<CategoryOptionGroupSet>
    created: string
    createdBy: User
    dataViewMaxOrganisationUnitLevel: number
    dataViewOrganisationUnits: Array<OrganisationUnit>
    disabled: boolean
    displayName: string
    education: string
    email: string
    employer: string
    externalAuth: boolean
    facebookMessenger: string
    favorite: boolean
    favorites: Array<string>
    firstName: string
    gender: string
    href: string
    id: string
    interests: string
    introduction: string
    invitation: boolean
    jobTitle: string
    languages: string
    lastCheckedInterpretations: string
    lastLogin: string
    lastUpdated: string
    lastUpdatedBy: User
    ldapId: string
    nationality: string
    openId: string
    organisationUnits: Array<OrganisationUnit>
    password: string
    passwordLastUpdated: string
    phoneNumber: string
    selfRegistered: boolean
    settings: UserSettings
    sharing: Sharing
    skype: string
    surname: string
    teiSearchOrganisationUnits: Array<OrganisationUnit>
    telegram: string
    translations: Array<Translation>
    twitter: string
    twoFactorEnabled: boolean
    user: User
    userCredentials: UserCredentialsDto
    userGroups: Array<UserGroup>
    userRoles: Array<UserRole>
    username: string
    welcomeMessage: string
    whatsApp: string
}

export type UserAccess = {
    access: string
    displayName: string
    id: string
}
// ...
export type UserGroupAccess = {
    access: string
    displayName: string
    id: string
}
// ...
export type UserSettings = {
    analysisDisplayProperty: UserSettings.analysisDisplayProperty
    dbLocale: string
    messageEmailNotification: boolean
    messageSmsNotification: boolean
    style: string
    trackerDashboardLayout: string
    uiLocale: string
}

export namespace UserSettings {
    export enum analysisDisplayProperty {
        NAME = 'NAME',
        SHORTNAME = 'SHORTNAME',
    }
}